app/backend/requirements.txt
flask
pdf2image
pytesseract
pillow
opencv-python-headless
PyMuPDF
numpy
boto3
requests
python-dotenv
docling
